p4i integration Integrate 242276 from 4.2 to main: Fixes:    QDataStream's streaming of qreals broke on platforms where qreal != double Task:     138732 RevBy:    Lars AutoTest: 242273 Details:  Ensure that all qreals that are written to the stream are cast 	  to double, and that we always read into doubles when reading. This 	  ensures that datastream output generated on one platform will be 	  readable on platforms that define qreal to be something different 	  than a double.
[git-p4: depot-paths = "//depot/qt/main/": change = 242277]

+++ src/corelib/tools/qpoint.cpp
@@ -671,7 +671,7 @@ QDebug operator<<(QDebug d, const QPointF &p)
 
 QDataStream &operator<<(QDataStream &s, const QPointF &p)
 {
-    s << p.x() << p.y();
+    s << double(p.x()) << double(p.y());
     return s;
 }
 
@@ -690,8 +690,8 @@ QDataStream &operator>>(QDataStream &s, QPointF &p)
     double x, y;
     s >> x;
     s >> y;
-    p.setX(x);
-    p.setY(y);
+    p.setX(qreal(x));
+    p.setY(qreal(y));
     return s;
 }
 #endif // QT_NO_DATASTREAM
